requirements.txt
# Picovoice stuff
pvporcupine
pvrecorder

# Vosk
vosk

PyYAML
python-dotenv